pyproject.toml
[tool.poetry]
name = "stackoverflow"
version = "0.1.0"
description = ""
authors = ["Ihor Horobets <ihor@horobets.me>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
djangorestframework = "^3.11.0"
django-rest-knox = "^4.1.0"
psycopg2-binary = "^2.8.5"
pillow = "^7.1.2"

[tool.poetry.dev-dependencies]
django-extensions = "^2.2.9"
pylint = "^2.5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
